requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
plotly>=5.18.0
openpyxl>=3.1.0
pyyaml>=6.0.0
anthropic>=0.18.0
cryptography>=41.0.0
google-generativeai>=0.3.0
Pillow>=10.0.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
pypdf>=4.0.0
